test(ui): seat four tests beside the functions they cover
Quick-close behavior tests join quick_test.go, review-header
formatting tests join diffrender_test.go.
