+++ internal/ui/diffrender_test.go
@@ -1,6 +1,7 @@
 package ui
 
 import (
+	"path/filepath"
 	"strings"
 	"testing"
 
@@ -141,5 +142,78 @@ func TestHLCacheEvicts(t *testing.T) {
 	}
 	if len(cache.entries) != highlightCacheCap {
 		t.Fatalf("cache size = %d", len(cache.entries))
+	}
+}
+
+// The review header always names the repo and its branch; in branch scope it
+// shows the base and the branch it diffs into.
+func TestReviewHeaderShowsRepoBranchAndBase(t *testing.T) {
+	m := buildModel(t)
+	if m.gitDrv == nil {
+		t.Skip("git not installed")
+	}
+	dir := gitRepoWithTwoChangedFiles(t)
+	openReviewOn(t, m, "hdr", dir)
+
+	header := m.viewDiffHeader("hdr")
+	if !strings.Contains(header, filepath.Base(dir)) {
+		t.Fatalf("header should name the repo, got %q", header)
+	}
+	if !strings.Contains(header, "main") {
+		t.Fatalf("header should show the branch, got %q", header)
+	}
+
+	for m.diff.scope != git.ScopeBranch {
+		m.drainCmds(t, m.cycleDiffScope())
+	}
+	header = m.viewDiffHeader("hdr")
+	if !strings.Contains(header, "→ main") {
+		t.Fatalf("branch scope header should show base → branch, got %q", header)
+	}
+}
+
+// The target pill in the header is the one place a reviewer figures out
+// what they are diffing into, so it has to read cleanly: the @<hash>
+// suffix BaseDesc carries internally is dropped, each changeable pill
+// wears its key, and an auto-detected target says so out loud while an
+// explicitly set one does not.
+func TestReviewHeaderTargetLabelCleanAndKeyed(t *testing.T) {
+	m := buildModel(t)
+	if m.gitDrv == nil {
+		t.Skip("git not installed")
+	}
+	dir := gitRepoWithSecondBranch(t)
+	openReviewOn(t, m, "hdr", dir)
+	sess, ok := m.diffSession()
+	if !ok {
+		t.Fatal("no diff session")
+	}
+	for m.diff.scope != git.ScopeBranch {
+		m.drainCmds(t, m.cycleDiffScope())
+	}
+
+	header := ansi.Strip(m.viewDiffHeader("hdr"))
+	if !strings.Contains(header, "B ") || !strings.Contains(header, "main") {
+		t.Fatalf("target pill should wear its B key, got %q", header)
+	}
+	if strings.Contains(header, "@") {
+		t.Fatalf("header should drop the @hash suffix from the target, got %q", header)
+	}
+	if !strings.Contains(header, "(auto)") {
+		t.Fatalf("auto-detected target should be marked, got %q", header)
+	}
+
+	if err := m.store.SetReviewBase(sess.ID, m.diff.repoSel, "feature"); err != nil {
+		t.Fatal(err)
+	}
+	m.diff.set.BaseOverride = "feature"
+	m.diff.set.BaseDesc = "feature@deadbee"
+	m.diff.set.Repo.Branch = "feature"
+	header = ansi.Strip(m.viewDiffHeader("hdr"))
+	if strings.Contains(header, "(auto)") {
+		t.Fatalf("explicit target should not be marked auto, got %q", header)
+	}
+	if !strings.Contains(header, "feature → feature") {
+		t.Fatalf("header should show the cleaned target → branch, got %q", header)
 	}
 }
