+++ internal/ui/settings_test.go
@@ -30,37 +30,6 @@ func TestDefaultSplitLayout(t *testing.T) {
 	}
 }
 
-func TestQuickCloseAfterSendDefaultsToStayingOpen(t *testing.T) {
-	m := buildModel(t)
-	if m.quickCloseAfterSend() {
-		t.Fatal("quick bar should stay open by default")
-	}
-	if err := m.store.SetSetting(quickCloseSetting, "close"); err != nil {
-		t.Fatal(err)
-	}
-	if !m.quickCloseAfterSend() {
-		t.Fatal("stored close choice should opt in")
-	}
-}
-
-func TestQuickPromptClosesAfterSendWhenEnabled(t *testing.T) {
-	m := buildModel(t)
-	createSession(t, m, "answer-me", t.TempDir(), "")
-	m.selectSessionRow(t, "answer-me")
-	if err := m.store.SetSetting(quickCloseSetting, "close"); err != nil {
-		t.Fatal(err)
-	}
-
-	m.openQuickMode()
-	m.quick.input.SetValue("carry on with the plan")
-	if _, _ = m.submitQuick(); m.errBar.text != "" {
-		t.Fatalf("send: %q", m.errBar.text)
-	}
-	if m.quick.active {
-		t.Fatal("quick mode should close after a send when the setting is on")
-	}
-}
-
 func TestSettingsTogglesQuickClose(t *testing.T) {
 	m := buildModel(t)
 	m.openSettings()
